Create Decode a Map.java
Decode a Map
Question
Implement a method that decodes a String to a corresponding Map. The String is composed of key-value pairs. Key is separated from value with an 'equals' character ('='). Pairs are separated with an 'ampersand' character ('&'). Example: key=value&key=value&....
Empty keys and values are allowed, but the equals sign must be present (e.g. "=value", "key=").
If the key or value is empty, empty String should be returned.
If the given String is empty, an empty Map should be returned.
If the given String is null, null should be returned.
If the given String has an invalid format, an IllegalArgumentException should be thrown.
Your implementation must implement the MapDecoder interface.
Sample Input: "one=1&two=2"
should return a Map containing {"one": "1", "two": "2"}
Coding
Your class must be named mapdecoder.MyMapDecoder